Add files via upload
This is code that was pieced together, and tested.   My data is generated by legacy Solartron SMART software which allows data to be exported to CSV and typically includes frequency, and 4 other paramaters although the original export contains more columns (temperature etc) which are never used in my case.  This is probably applicable to other generated data not just Solartron.

The code is a bit more streamlined than the other files which I might update now with new knowledge.  This is all very basic stuff, but useful to a lot of people hopefully.